+++ ICD.Profound.ConnectPRO/Themes/UserInterface/IPresenters/VideoConference/IVtcKeypadPresenter.cs
@@ -3,8 +3,8 @@ using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.VideoConference;
 
 namespace ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.VideoConference
 {
-	public interface IVtcKeyboardPresenter : IPresenter<IVtcKeyboardView>
+	public interface IVtcKeypadPresenter : IPresenter<IVtcKeypadView>
 	{
-		event EventHandler OnKeypadButtonPressed; 
+		event EventHandler OnKeyboardButtonPressed;
 	}
 }